% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_counter.R
\name{add_counter}
\alias{add_counter}
\title{Add Counter}
\usage{
add_counter(df, col, reference_col = col, prefix = "other")
}
\arguments{
\item{df}{dataframe}

\item{col}{character specifying the column selection}

\item{reference_col}{character showing the reference column.  The default is
'col'}

\item{prefix}{character specifying the prefix that belongs to counter.  The
default is 'other'}
}
\value{
dataframe with 'col' modified.
}
\description{
This function identifies all NAs in a column and converts it to a character
with a counter, the default value being 'other_{counter}'.  The counter is
based on a reference column of choice, the default being the original
column.  It is not intended to use the original column as the reference but
if that's chosen all NA's will show '_1'.
}
