% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_order.R
\name{apply_col_order}
\alias{apply_col_order}
\title{Apply Column Order}
\usage{
apply_col_order(df, col_order = sort(names(df)))
}
\arguments{
\item{df}{dataframe}

\item{col_order}{character vector}
}
\value{
dataframe ordered by column
}
\description{
Apply Column Order
}
