% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_order.R
\name{apply_row_order}
\alias{apply_row_order}
\title{Apply Row Order}
\usage{
apply_row_order(df, cols = names(df), reverse_cols = NULL)
}
\arguments{
\item{df}{dataframe}

\item{cols}{character vector}

\item{reverse_cols}{character vector or NULL (default)}
}
\value{
dataframe
}
\description{
Apply Row Order
}
