% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate Difference}
\usage{
calc_diff(x, y)
}
\arguments{
\item{x}{numeric value, or vector}

\item{y}{numeric value, or vector}
}
\value{
numeric value, or vector
}
\description{
This function calculates the difference between two values while ensuring
there are no adverse results such as NA, NAN, or Inf.
}
