% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_margin}
\alias{calc_margin}
\title{Calculate Margin}
\usage{
calc_margin(x, y)
}
\arguments{
\item{x}{numeric value, or vector}

\item{y}{numeric value, or vector}
}
\value{
numeric value, or vector
}
\description{
This function calculates the margin, or relative size, of a value compared to
another while ensuring there are no adverse results such as NA, NAN, or Inf.
It also attempts to avoid the issue of returning infinitely small, non-zero,
values by rounding each value to two decimals first.
}
