% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols.R
\name{date_cols}
\alias{date_cols}
\title{Date Columns}
\usage{
date_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
a character vector of columns that correspond to all date columns
}
\description{
Date Columns
}
