% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{depth}
\alias{depth}
\title{Depth}
\usage{
depth(x, d = 0)
}
\arguments{
\item{x}{R object}

\item{d}{numeric - starting depth.  Default = 0.}
}
\value{
numeric - showing the number of levels of the list.  0 is an object
 that is not a list.
}
\description{
This allows for finding the depth of a nested list
}
