% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_null.R
\name{discard_null}
\alias{discard_null}
\title{Discard Null}
\usage{
discard_null(x)
}
\arguments{
\item{x}{list}
}
\value{
a list without NULL values
}
\description{
This is a recursive function that removes all NULL values from a list.
}
