% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{eval_sym}
\alias{eval_sym}
\title{Evaluate Symbol}
\usage{
eval_sym(x)
}
\arguments{
\item{x}{characters - usually representing columns in a dataframe}
}
\description{
This functions purpose is to eliminate problems with global variable bindings
by using characters instead of symbols.  This function takes the character,
converts it to a symbol and evaluates.
}
