% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataframe.R
\name{fmt_percent_by_col}
\alias{fmt_percent_by_col}
\title{Format Dataframe Percent (Column)}
\usage{
fmt_percent_by_col(df, cols = numeric_cols(df))
}
\arguments{
\item{df}{dataframe}

\item{cols}{numeric}
}
\value{
dataframe
}
\description{
Format Dataframe Percent (Column)
}
