% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataframe.R
\name{fmt_percent_by_row_and_col}
\alias{fmt_percent_by_row_and_col}
\title{Format Dataframe Percent (Row and Column)}
\usage{
fmt_percent_by_row_and_col(df, rows, cols = numeric_cols(df))
}
\arguments{
\item{df}{dataframe}

\item{rows}{numeric}

\item{cols}{numeric}
}
\value{
dataframe
}
\description{
Format Dataframe Percent (Row and Column)
}
