% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataframe.R
\name{fmt_value_by_row}
\alias{fmt_value_by_row}
\title{Format Dataframe Values (Row)}
\usage{
fmt_value_by_row(df, rows = 1:nrow(df))
}
\arguments{
\item{df}{dataframe}

\item{rows}{numeric}
}
\value{
dataframe
}
\description{
Format Dataframe Values (Row)
}
