% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataframe.R
\name{fmt_value_by_row_and_col}
\alias{fmt_value_by_row_and_col}
\title{Format Dataframe Values (Row and Column)}
\usage{
fmt_value_by_row_and_col(df, rows = 1:nrow(df),
  cols = numeric_cols(df))
}
\arguments{
\item{df}{dataframe}

\item{rows}{numeric}

\item{cols}{numeric}
}
\value{
dataframe
}
\description{
Format Dataframe Values (Row and Column)
}
