% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.R
\name{grep_0}
\alias{grep_0}
\title{Grep Zero}
\usage{
grep_0(pattern, x, value = TRUE, invert = FALSE)
}
\arguments{
\item{pattern}{character}

\item{x}{character}

\item{value}{logical}

\item{invert}{logical}
}
\value{
character
}
\description{
A wrapper function for 'grep' that ignores case and returns values (default).
}
