% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_and_summarize.R
\name{group_and_summarize}
\alias{group_and_summarize}
\title{Group and Summarize}
\usage{
group_and_summarize(df, groups = non_numeric_cols(df),
  summary_cols = numeric_cols(df), negate = FALSE, .f = sum_na)
}
\arguments{
\item{df}{data.frame}

\item{groups}{character}

\item{summary_cols}{character vector of columns to summarize.  This
defaults to all numeric columns, but could be switched to other columns
if the desire, for instance, is to concatenate all characters.}

\item{negate}{logical}

\item{.f}{function to be applied.
Default is 'sum_na', or, sum(x, na.rm = TRUE).}
}
\value{
data.frame
}
\description{
Group and Summarize
}
