% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inter_cols}
\alias{inter_cols}
\title{Intersecting Columns}
\usage{
inter_cols(df1, df2)
}
\arguments{
\item{df1}{dataframe1}

\item{df2}{dataframe2}
}
\value{
This returns a character vector listing the names of the intersecting
  columns of the two dataframes listed above.
}
\description{
Intersecting Columns
}
