% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_difference.R
\name{make_difference}
\alias{make_difference}
\title{Make Data - Difference}
\usage{
make_difference(df, include_growth = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{include_growth}{logical - Do you want the growth between columns
along with the difference?}
}
\value{
dataframe
}
\description{
This will produce a dataframe with additional columns showing the
difference between each numeric column, subsequent to the first, and the
first numeric column.
}
