% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{map_date}
\alias{map_date}
\title{Map Dates}
\usage{
map_date(values, style = "monthly")
}
\arguments{
\item{values}{character, usually a vector of dataframe names}

\item{style}{character, either 'monthly' or 'weekly'}
}
\value{
a vector of equal length as the ones provided with the values
 being transformed into a a readable format
}
\description{
Map Dates
}
