% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{map_funct}
\alias{map_funct}
\title{Map Function}
\usage{
map_funct(values, .f)
}
\arguments{
\item{values}{vector to be altered}

\item{.f}{function to apply to the provided vector}
}
\value{
vector of equal length
}
\description{
This is similar to mapping values except instead of providing two vectors
of equal length, you provide one vector and function.  The function
provided will be applied to the vector, and any values that cannot be
coerced will be returned as-is.
}
