% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{map_snake}
\alias{map_snake}
\title{Map Values - Snakecase}
\usage{
map_snake(values)
}
\arguments{
\item{values}{character vector}
}
\value{
character vector of equal length that transforms the original
 vector to snakecase
}
\description{
Predefined version of 'map_funct' that uses snakecase::to_snake_case
}
