% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols.R
\name{non_numeric_cols}
\alias{non_numeric_cols}
\title{Non-Numeric Columns}
\usage{
non_numeric_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
a character vector of columns that correspond to all
 non-numeric columns
}
\description{
Non-Numeric Columns
}
