% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols.R
\name{numeric_cols}
\alias{numeric_cols}
\title{Numeric Columns}
\usage{
numeric_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
a character vector of columns that correspond to all numeric columns
}
\description{
Numeric Columns
}
