% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition}
\usage{
partition(df, cols, drop = TRUE)
}
\arguments{
\item{df}{data}

\item{cols}{character vector of columns to split the dataframe.}

\item{drop}{logical.  When splitting dataframes should the factors not in
the dataset be dropped?  default = TRUE.}
}
\value{
recursive list
}
\description{
This creates a recursive list by splitting a dataframe sequencially by
the provided columns.
}
