% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste0_n}
\alias{paste0_n}
\title{Paste - collapse ('new_line')}
\usage{
paste0_n(...)
}
\arguments{
\item{...}{elements passed to 'paste'}
}
\value{
A character vector of the concatenated values
}
\description{
Paste - collapse ('new_line')
}
