% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_n}
\alias{paste_n}
\title{Paste - sep ('new_line')}
\usage{
paste_n(...)
}
\arguments{
\item{...}{elements passed to 'paste'}
}
\value{
A character vector of the concatenated values
}
\description{
Paste - sep ('new_line')
}
