% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{pmap_values}
\alias{pmap_values}
\title{Map Values - Partial}
\usage{
pmap_values(values, map, replacement = NULL)
}
\arguments{
\item{values}{vector of values to match/replace}

\item{map}{named character vector to pass to 'grep_v'}

\item{replacement}{replacement value.  If NULL (default) the original value
is returned.}
}
\value{
vector of same length as the original value.
}
\description{
Map Values - Partial
}
