% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{repl_na}
\alias{repl_na}
\title{Replace NA's}
\usage{
repl_na(x, replacement = 0)
}
\arguments{
\item{x}{vector, usually numeric.}

\item{replacement}{value, usually character or numeric, that is used as the
replacement for all NA, NAN, or Inf values found.  default = 0.}
}
\value{
vector
}
\description{
This will replace all na, nan, inf values in a vector
}
