% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shear.R
\name{shear}
\alias{shear}
\title{Shear}
\usage{
shear(df, ...)
}
\arguments{
\item{df}{dataframe to be filtered}

\item{...}{a named set of values where the names are the columns to be
filtered and the values are the filtering criteria.}
}
\value{
dataframe
}
\description{
This allows for filtering dataframes and then eliminating the filtered
column.  It's a quick way to simplify data.  It's also a good way to
simplify a range of datasets that may not have similar features because if
the columns to be filtered do no exist in the dataframe the original
dataframe is returned.
}
